fix a few casting issues allowing maps such as warpa, etc. to load (probably not done here)
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@3028 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/common/gl_q2bsp.c
@@ -1660,14 +1660,14 @@ qboolean CMod_LoadLeafs (lump_t *l)
 		}
 
 		out->contents = LittleLong (in->contents);
-		out->cluster = LittleShort (in->cluster);
+		out->cluster = (unsigned short)LittleShort (in->cluster);
 		out->area = LittleShort (in->area);
-		out->firstleafbrush = LittleShort (in->firstleafbrush);
-		out->numleafbrushes = LittleShort (in->numleafbrushes);
+		out->firstleafbrush = (unsigned short)LittleShort (in->firstleafbrush);
+		out->numleafbrushes = (unsigned short)LittleShort (in->numleafbrushes);
 
 		out->firstmarksurface = loadmodel->marksurfaces +
-			LittleShort(in->firstleafface);
-		out->nummarksurfaces = LittleShort(in->numleaffaces);
+			(unsigned short)LittleShort(in->firstleafface);
+		out->nummarksurfaces = (unsigned short)LittleShort(in->numleaffaces);
 
 		if (out->cluster >= numclusters)
 			numclusters = out->cluster + 1;
